pyproject.toml
[project]
name = "llm-battle-royale"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "google-genai>=1.66.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.41.0",
]
